Rename keyring pkg to credentials
