+++ cmd/keyring/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/swampapp/swamp/internal/keyring"
+	"github.com/swampapp/swamp/internal/credentials"
 )
 
 func main() {
@@ -15,6 +15,6 @@ func main() {
 	repo := os.Args[1]
 	fmt.Println(repo)
 	os.Exit(0)
-	settings := keyring.New(repo)
+	settings := credentials.New(repo)
 	settings.Delete()
 }
