+++ internal/ui/assistant/assistant.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gotk3/gotk3/gtk"
 	"github.com/rubiojr/rapi"
 	"github.com/swampapp/swamp/internal/config"
-	"github.com/swampapp/swamp/internal/keyring"
+	"github.com/swampapp/swamp/internal/credentials"
 	"github.com/swampapp/swamp/internal/logger"
 	"github.com/swampapp/swamp/internal/ui/component"
 )
@@ -47,7 +47,7 @@ func New() *Assistant {
 		}
 	})
 
-	var rs *keyring.Keyring
+	var rs *credentials.Credentials
 	page2 := a.GladeWidget("assistantPage2").(*gtk.Box)
 	btn := a.GladeWidget("testSettingsBTN").(*gtk.Button)
 	var repoID, repoName, ruri, rpass, var1, var2 string
@@ -86,7 +86,7 @@ func New() *Assistant {
 			} else {
 				glib.IdleAdd(func() {
 					repoID = repo.Config().ID
-					rs = keyring.New(repoID)
+					rs = credentials.New(repoID)
 					rs.Password = rpass
 					rs.Repository = ruri
 					rs.Var1 = var1
